+++ src/client/components/evaluate/ProjectEvaluateMetricDetail.tsx
@@ -1,12 +1,18 @@
 /** @jsx jsx */
 import { Box, Button, Flex, jsx, ThemeUIStyleObject, Heading } from "theme-ui";
-
-import { EvaluateMetric } from "../../../shared/entities";
+import {
+  EvaluateMetric,
+  IProject,
+  IStaticMetadata,
+  RegionLookupProperties
+} from "../../../shared/entities";
 import Icon from "../Icon";
 import { DistrictsGeoJSON } from "../../types";
 import store from "../../store";
 import { selectEvaluationMetric } from "../../actions/districtDrawing";
 import CompactnessMetricDetail from "./detail/Compactness";
+import CountySplitMetricDetail from "./detail/CountySplit";
+import { Resource } from "../../resource";
 
 const style: ThemeUIStyleObject = {
   td: {
@@ -52,53 +58,77 @@ const style: ThemeUIStyleObject = {
   },
   metricText: {
     fontSize: "10",
+    ml: "50px",
     minHeight: "100px",
     fontWeight: "400"
+  },
+  metricDescription: {
+    fontWeight: "500",
+    color: "gray.8",
+    minHeight: "200px"
   }
 };
 
 const ProjectEvaluateMetricDetail = ({
   geojson,
-  metric
+  metric,
+  project,
+  regionProperties,
+  geoLevel,
+  staticMetadata
 }: {
   readonly geojson?: DistrictsGeoJSON;
   readonly metric: EvaluateMetric;
+  readonly project?: IProject;
+  readonly regionProperties: Resource<readonly RegionLookupProperties[]>;
+  readonly geoLevel: string;
+  readonly staticMetadata?: IStaticMetadata;
 }) => {
   return (
-    <Flex sx={{ flexDirection: "column", height: "100vh" }}>
-      <Flex sx={style.header} className="evaluate-metric-header">
-        <Box sx={{ display: "block" }}>
-          <Button onClick={() => store.dispatch(selectEvaluationMetric(undefined))}>
-            <Icon name="chevron-left" /> Back to summary
-          </Button>
-        </Box>
-      </Flex>
-      <Box sx={{ display: "block", m: "20px", borderBottom: "1px solid gray" }}>
-        <Heading as="h2" sx={{ variant: "text.h4", m: "0" }}>
-          <Icon name={"check"} /> {metric.name}
-        </Heading>
-        <Flex sx={style.metricText}>{metric.longText || "Lorem ipsum lorem ipsum"}</Flex>
-      </Box>
-      {metric && "type" in metric && metric.key === "compactness" ? (
-        <CompactnessMetricDetail metric={metric} geojson={geojson} />
-      ) : (
-        <Box sx={{ ml: "20px" }}>
-          <Heading as="h4" sx={{ variant: "text.h4", display: "block" }}>
-            7 / 10 districts {metric.description}
+    <Box>
+      <Flex sx={{ flexDirection: "column", height: "100vh" }}>
+        <Flex sx={style.header} className="evaluate-metric-header">
+          <Box sx={{ display: "block" }}>
+            <Button onClick={() => store.dispatch(selectEvaluationMetric(undefined))}>
+              <Icon name="chevron-left" /> Back to summary
+            </Button>
+          </Box>
+        </Flex>
+        <Box sx={{ display: "block", m: "20px", borderBottom: "1px solid gray" }}>
+          <Heading as="h2" sx={{ variant: "text.h4", m: "0", textTransform: "capitalize" }}>
+            <Icon name={"check"} /> {metric.name}
           </Heading>
-          <Flex sx={{ flexDirection: "column" }}>
-            {geojson?.features.map(
-              d =>
-                d.properties.contiguity !== "" && (
-                  <Box sx={{ display: "block" }} key={d.id}>
-                    {d.id}
-                  </Box>
-                )
-            )}
-          </Flex>
+          <Flex sx={style.metricDescription}>{metric.longText || "Lorem ipsum lorem ipsum"} </Flex>
         </Box>
-      )}
-    </Flex>
+        {metric && "type" in metric && metric.key === "countySplits" ? (
+          <CountySplitMetricDetail
+            project={project}
+            metric={metric}
+            geoLevel={geoLevel}
+            regionProperties={regionProperties}
+            staticMetadata={staticMetadata}
+          />
+        ) : metric && "type" in metric && metric.key === "compactness" ? (
+          <CompactnessMetricDetail metric={metric} geojson={geojson} />
+        ) : (
+          <Box sx={{ ml: "20px" }}>
+            <Heading as="h4" sx={{ variant: "text.h4", display: "block" }}>
+              7 / 10 districts {metric.description}
+            </Heading>
+            <Flex sx={{ flexDirection: "column" }}>
+              {geojson?.features.map(
+                d =>
+                  d.properties.contiguity !== "" && (
+                    <Box sx={{ display: "block" }} key={d.id}>
+                      {d.id}
+                    </Box>
+                  )
+              )}
+            </Flex>
+          </Box>
+        )}
+      </Flex>
+    </Box>
   );
 };