+++ src/server/src/region-configs/region-configs.module.ts
@@ -3,11 +3,12 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import { RegionConfigsController } from "./controllers/region-configs.controller";
 import { RegionConfig } from "./entities/region-config.entity";
 import { RegionConfigsService } from "./services/region-configs.service";
+import { TopologyService } from "../districts/services/topology.service";
 
 @Module({
   imports: [TypeOrmModule.forFeature([RegionConfig])],
   controllers: [RegionConfigsController],
-  providers: [RegionConfigsService],
+  providers: [RegionConfigsService, TopologyService],
   exports: [RegionConfigsService, TypeOrmModule]
 })
 export class RegionConfigsModule {}